Jira # 159
There should be decrease and increase buttons as well as a text box to edit the quantity in the cart page.

There should be a listing with the subtotal, tax, and total price.

There should be a button that directs you to the checkout.

Once the user has clicked checkout, create an order in the database containing their corresponding product id’s and quantity.